+++ packages/renderer/src/components/Modals/PublishProject/AlternativeServers/index.ts
@@ -1,0 +1,1 @@
+export { AlternativeServers } from './component';
