+++ packages/renderer/src/components/Modals/PublishProject/Deploy/index.ts
@@ -1,0 +1,1 @@
+export { Deploy } from './component';
