+++ packages/renderer/src/components/Modals/PublishProject/Deploy/styles.css
@@ -1,0 +1,3 @@
+.Deploy {
+  color: red;
+}
