+++ packages/renderer/src/components/Modals/PublishProject/Initial/index.ts
@@ -1,0 +1,1 @@
+export { Initial } from './component';
