+++ packages/renderer/src/components/Modals/PublishProject/index.ts
@@ -1,3 +1,3 @@
 import { PublishProject } from './component';
-import type { StepValue } from './types';
-export { PublishProject, StepValue };
+import type { TargetValue } from './types';
+export { PublishProject, TargetValue as StepValue };
