+++ packages/renderer/src/components/Modals/PublishProject/types.ts
@@ -4,13 +4,18 @@ export type Props = {
   open: boolean;
   project: Project;
   onClose: () => void;
-  onSubmit: (value: StepValue) => void;
+  onTarget: (value: TargetValue) => void;
 };
 
-export type Step = 'initial' | 'alternative-servers' | 'publish-to-world' | 'publish-to-land';
+export type Step =
+  | 'initial'
+  | 'alternative-servers'
+  | 'publish-to-world'
+  | 'publish-to-land'
+  | 'deploy';
 export type InitialTarget = 'worlds' | 'land';
 export type AlternativeTarget = 'test' | 'custom';
 
 export type Target = InitialTarget | AlternativeTarget;
-export type StepValue = { target: Target; value?: string };
-export type StepProps = { onClick: (value: StepValue) => void };
+export type TargetValue = { target: Target; value?: string };
+export type TargetProps = { onTarget: (value: TargetValue) => void };
